Add router and organize the folders
